Implement requested changes
